chore(scripts): share teamclaw-introspect sidecar prep in ensure-introspect-sidecar
Extract build-and-install logic for rust-cli and tauri-cli; skip when CI is set.
